Add files via upload
This set of R files is what Iuse to execute R/earth MARS regression on property date from pro.mlslistings.com, from reading the MLS data, preprocessing it, running  1 or 2 degree earth regeression, producing the model as code and grpahs, associated statistical graphs and information, to creating a an Excel spreadsheet with selected comps to upload to the Alamode Worksheet.  In this case the upload to Alamode is for URAR forms.  But you can create uploads for other types of forms such as 1025, HOAs and so ojn.

Signed-off-by: William Bert Craytor <69237297+bert-craytor@users.noreply.github.com>